# Improve performance by reducing global allocations.