--- a/main.cpp
+++ b/main.cpp
@@ -697,6 +697,298 @@ static ExecutionModel stage_to_execution_model(const std::string &stage)
 		SPIRV_CROSS_THROW("Invalid stage.");
 }
 
+static string compile_iteration(const CLIArguments &args, vector<uint32_t> spirv_file)
+{
+	Parser spirv_parser(move(spirv_file));
+	spirv_parser.parse();
+
+	unique_ptr<CompilerGLSL> compiler;
+	bool combined_image_samplers = false;
+	bool build_dummy_sampler = false;
+
+	if (args.cpp)
+	{
+		compiler.reset(new CompilerCPP(move(spirv_parser.get_parsed_ir())));
+		if (args.cpp_interface_name)
+			static_cast<CompilerCPP *>(compiler.get())->set_interface_name(args.cpp_interface_name);
+	}
+	else if (args.msl)
+	{
+		compiler.reset(new CompilerMSL(move(spirv_parser.get_parsed_ir())));
+
+		auto *msl_comp = static_cast<CompilerMSL *>(compiler.get());
+		auto msl_opts = msl_comp->get_msl_options();
+		if (args.set_msl_version)
+			msl_opts.msl_version = args.msl_version;
+		msl_opts.capture_output_to_buffer = args.msl_capture_output_to_buffer;
+		msl_opts.swizzle_texture_samples = args.msl_swizzle_texture_samples;
+		if (args.msl_ios)
+			msl_opts.platform = CompilerMSL::Options::iOS;
+		msl_opts.pad_fragment_output_components = args.msl_pad_fragment_output;
+		msl_opts.tess_domain_origin_lower_left = args.msl_domain_lower_left;
+		msl_opts.argument_buffers = args.msl_argument_buffers;
+		msl_comp->set_msl_options(msl_opts);
+		for (auto &v : args.msl_discrete_descriptor_sets)
+			msl_comp->add_discrete_descriptor_set(v);
+	}
+	else if (args.hlsl)
+		compiler.reset(new CompilerHLSL(move(spirv_parser.get_parsed_ir())));
+	else
+	{
+		combined_image_samplers = !args.vulkan_semantics;
+		if (!args.vulkan_semantics)
+			build_dummy_sampler = true;
+		compiler.reset(new CompilerGLSL(move(spirv_parser.get_parsed_ir())));
+	}
+
+	if (!args.variable_type_remaps.empty())
+	{
+		auto remap_cb = [&](const SPIRType &, const string &name, string &out) -> void {
+			for (const VariableTypeRemap &remap : args.variable_type_remaps)
+				if (name == remap.variable_name)
+					out = remap.new_variable_type;
+		};
+
+		compiler->set_variable_type_remap_callback(move(remap_cb));
+	}
+
+	for (auto &rename : args.entry_point_rename)
+		compiler->rename_entry_point(rename.old_name, rename.new_name, rename.execution_model);
+
+	auto entry_points = compiler->get_entry_points_and_stages();
+	auto entry_point = args.entry;
+	ExecutionModel model = ExecutionModelMax;
+
+	if (!args.entry_stage.empty())
+	{
+		model = stage_to_execution_model(args.entry_stage);
+		if (entry_point.empty())
+		{
+			// Just use the first entry point with this stage.
+			for (auto &e : entry_points)
+			{
+				if (e.execution_model == model)
+				{
+					entry_point = e.name;
+					break;
+				}
+			}
+
+			if (entry_point.empty())
+			{
+				fprintf(stderr, "Could not find an entry point with stage: %s\n", args.entry_stage.c_str());
+				exit(EXIT_FAILURE);
+			}
+		}
+		else
+		{
+			// Make sure both stage and name exists.
+			bool exists = false;
+			for (auto &e : entry_points)
+			{
+				if (e.execution_model == model && e.name == entry_point)
+				{
+					exists = true;
+					break;
+				}
+			}
+
+			if (!exists)
+			{
+				fprintf(stderr, "Could not find an entry point %s with stage: %s\n", entry_point.c_str(),
+				        args.entry_stage.c_str());
+				exit(EXIT_FAILURE);
+			}
+		}
+	}
+	else if (!entry_point.empty())
+	{
+		// Make sure there is just one entry point with this name, or the stage
+		// is ambiguous.
+		uint32_t stage_count = 0;
+		for (auto &e : entry_points)
+		{
+			if (e.name == entry_point)
+			{
+				stage_count++;
+				model = e.execution_model;
+			}
+		}
+
+		if (stage_count == 0)
+		{
+			fprintf(stderr, "There is no entry point with name: %s\n", entry_point.c_str());
+			exit(EXIT_FAILURE);
+		}
+		else if (stage_count > 1)
+		{
+			fprintf(stderr, "There is more than one entry point with name: %s. Use --stage.\n", entry_point.c_str());
+			exit(EXIT_FAILURE);
+		}
+	}
+
+	if (!entry_point.empty())
+		compiler->set_entry_point(entry_point, model);
+
+	if (!args.set_version && !compiler->get_common_options().version)
+	{
+		fprintf(stderr, "Didn't specify GLSL version and SPIR-V did not specify language.\n");
+		print_help();
+		exit(EXIT_FAILURE);
+	}
+
+	CompilerGLSL::Options opts = compiler->get_common_options();
+	if (args.set_version)
+		opts.version = args.version;
+	if (args.set_es)
+		opts.es = args.es;
+	opts.force_temporary = args.force_temporary;
+	opts.separate_shader_objects = args.sso;
+	opts.flatten_multidimensional_arrays = args.flatten_multidimensional_arrays;
+	opts.enable_420pack_extension = args.use_420pack_extension;
+	opts.vulkan_semantics = args.vulkan_semantics;
+	opts.vertex.fixup_clipspace = args.fixup;
+	opts.vertex.flip_vert_y = args.yflip;
+	opts.vertex.support_nonzero_base_instance = args.support_nonzero_baseinstance;
+	opts.emit_push_constant_as_uniform_buffer = args.glsl_emit_push_constant_as_ubo;
+	compiler->set_common_options(opts);
+
+	// Set HLSL specific options.
+	if (args.hlsl)
+	{
+		auto *hlsl = static_cast<CompilerHLSL *>(compiler.get());
+		auto hlsl_opts = hlsl->get_hlsl_options();
+		if (args.set_shader_model)
+		{
+			if (args.shader_model < 30)
+			{
+				fprintf(stderr, "Shader model earlier than 30 (3.0) not supported.\n");
+				exit(EXIT_FAILURE);
+			}
+
+			hlsl_opts.shader_model = args.shader_model;
+		}
+
+		if (args.hlsl_compat)
+		{
+			// Enable all compat options.
+			hlsl_opts.point_size_compat = true;
+			hlsl_opts.point_coord_compat = true;
+		}
+
+		if (hlsl_opts.shader_model <= 30)
+		{
+			combined_image_samplers = true;
+			build_dummy_sampler = true;
+		}
+
+		hlsl_opts.support_nonzero_base_vertex_base_instance = args.hlsl_support_nonzero_base;
+		hlsl->set_hlsl_options(hlsl_opts);
+	}
+
+	if (build_dummy_sampler)
+	{
+		uint32_t sampler = compiler->build_dummy_sampler_for_combined_images();
+		if (sampler != 0)
+		{
+			// Set some defaults to make validation happy.
+			compiler->set_decoration(sampler, DecorationDescriptorSet, 0);
+			compiler->set_decoration(sampler, DecorationBinding, 0);
+		}
+	}
+
+	ShaderResources res;
+	if (args.remove_unused)
+	{
+		auto active = compiler->get_active_interface_variables();
+		res = compiler->get_shader_resources(active);
+		compiler->set_enabled_interface_variables(move(active));
+	}
+	else
+		res = compiler->get_shader_resources();
+
+	if (args.flatten_ubo)
+	{
+		for (auto &ubo : res.uniform_buffers)
+			compiler->flatten_buffer_block(ubo.id);
+		for (auto &ubo : res.push_constant_buffers)
+			compiler->flatten_buffer_block(ubo.id);
+	}
+
+	auto pls_inputs = remap_pls(args.pls_in, res.stage_inputs, &res.subpass_inputs);
+	auto pls_outputs = remap_pls(args.pls_out, res.stage_outputs, nullptr);
+	compiler->remap_pixel_local_storage(move(pls_inputs), move(pls_outputs));
+
+	for (auto &ext : args.extensions)
+		compiler->require_extension(ext);
+
+	for (auto &remap : args.remaps)
+	{
+		if (remap_generic(*compiler, res.stage_inputs, remap))
+			continue;
+		if (remap_generic(*compiler, res.stage_outputs, remap))
+			continue;
+		if (remap_generic(*compiler, res.subpass_inputs, remap))
+			continue;
+	}
+
+	for (auto &rename : args.interface_variable_renames)
+	{
+		if (rename.storageClass == StorageClassInput)
+			spirv_cross_util::rename_interface_variable(*compiler, res.stage_inputs, rename.location,
+			                                            rename.variable_name);
+		else if (rename.storageClass == StorageClassOutput)
+			spirv_cross_util::rename_interface_variable(*compiler, res.stage_outputs, rename.location,
+			                                            rename.variable_name);
+		else
+		{
+			fprintf(stderr, "error at --rename-interface-variable <in|out> ...\n");
+			exit(EXIT_FAILURE);
+		}
+	}
+
+	if (args.dump_resources)
+	{
+		print_resources(*compiler, res);
+		print_push_constant_resources(*compiler, res.push_constant_buffers);
+		print_spec_constants(*compiler);
+		print_capabilities_and_extensions(*compiler);
+	}
+
+	if (combined_image_samplers)
+	{
+		compiler->build_combined_image_samplers();
+		if (args.combined_samplers_inherit_bindings)
+			spirv_cross_util::inherit_combined_sampler_bindings(*compiler);
+
+		// Give the remapped combined samplers new names.
+		for (auto &remap : compiler->get_combined_image_samplers())
+		{
+			compiler->set_name(remap.combined_id, join("SPIRV_Cross_Combined", compiler->get_name(remap.image_id),
+			                                           compiler->get_name(remap.sampler_id)));
+		}
+	}
+
+	if (args.hlsl)
+	{
+		auto *hlsl_compiler = static_cast<CompilerHLSL *>(compiler.get());
+		uint32_t new_builtin = hlsl_compiler->remap_num_workgroups_builtin();
+		if (new_builtin)
+		{
+			hlsl_compiler->set_decoration(new_builtin, DecorationDescriptorSet, 0);
+			hlsl_compiler->set_decoration(new_builtin, DecorationBinding, 0);
+		}
+	}
+
+	if (args.hlsl)
+	{
+		for (auto &remap : args.hlsl_attr_remap)
+			static_cast<CompilerHLSL *>(compiler.get())->add_vertex_attribute_remap(remap);
+	}
+
+	return compiler->compile();
+}
+
 static int main_inner(int argc, char *argv[])
 {
 	CLIArguments args;
@@ -819,13 +1111,9 @@ static int main_inner(int argc, char *argv[])
 
 	CLIParser parser{ move(cbs), argc - 1, argv + 1 };
 	if (!parser.parse())
-	{
 		return EXIT_FAILURE;
-	}
 	else if (parser.ended_state)
-	{
 		return EXIT_SUCCESS;
-	}
 
 	if (!args.input)
 	{
@@ -837,13 +1125,13 @@ static int main_inner(int argc, char *argv[])
 	auto spirv_file = read_spirv_file(args.input);
 	if (spirv_file.empty())
 		return EXIT_FAILURE;
-	Parser spirv_parser(move(spirv_file));
-
-	spirv_parser.parse();
 
 	// Special case reflection because it has little to do with the path followed by code-outputting compilers
 	if (!args.reflect.empty())
 	{
+		Parser spirv_parser(move(spirv_file));
+		spirv_parser.parse();
+
 		CompilerReflection compiler(move(spirv_parser.get_parsed_ir()));
 		compiler.set_format(args.reflect);
 		auto json = compiler.compile();
@@ -854,300 +1142,20 @@ static int main_inner(int argc, char *argv[])
 		return EXIT_SUCCESS;
 	}
 
-	unique_ptr<CompilerGLSL> compiler;
-	bool combined_image_samplers = false;
-	bool build_dummy_sampler = false;
+	string compiled_output;
 
-	if (args.cpp)
-	{
-		compiler.reset(new CompilerCPP(move(spirv_parser.get_parsed_ir())));
-		if (args.cpp_interface_name)
-			static_cast<CompilerCPP *>(compiler.get())->set_interface_name(args.cpp_interface_name);
-	}
-	else if (args.msl)
-	{
-		compiler.reset(new CompilerMSL(move(spirv_parser.get_parsed_ir())));
-
-		auto *msl_comp = static_cast<CompilerMSL *>(compiler.get());
-		auto msl_opts = msl_comp->get_msl_options();
-		if (args.set_msl_version)
-			msl_opts.msl_version = args.msl_version;
-		msl_opts.capture_output_to_buffer = args.msl_capture_output_to_buffer;
-		msl_opts.swizzle_texture_samples = args.msl_swizzle_texture_samples;
-		if (args.msl_ios)
-			msl_opts.platform = CompilerMSL::Options::iOS;
-		msl_opts.pad_fragment_output_components = args.msl_pad_fragment_output;
-		msl_opts.tess_domain_origin_lower_left = args.msl_domain_lower_left;
-		msl_opts.argument_buffers = args.msl_argument_buffers;
-		msl_comp->set_msl_options(msl_opts);
-		for (auto &v : args.msl_discrete_descriptor_sets)
-			msl_comp->add_discrete_descriptor_set(v);
-	}
-	else if (args.hlsl)
-		compiler.reset(new CompilerHLSL(move(spirv_parser.get_parsed_ir())));
+	if (args.iterations == 1)
+		compiled_output = compile_iteration(args, move(spirv_file));
 	else
 	{
-		combined_image_samplers = !args.vulkan_semantics;
-		if (!args.vulkan_semantics)
-			build_dummy_sampler = true;
-		compiler.reset(new CompilerGLSL(move(spirv_parser.get_parsed_ir())));
-	}
-
-	if (!args.variable_type_remaps.empty())
-	{
-		auto remap_cb = [&](const SPIRType &, const string &name, string &out) -> void {
-			for (const VariableTypeRemap &remap : args.variable_type_remaps)
-				if (name == remap.variable_name)
-					out = remap.new_variable_type;
-		};
-
-		compiler->set_variable_type_remap_callback(move(remap_cb));
-	}
-
-	for (auto &rename : args.entry_point_rename)
-		compiler->rename_entry_point(rename.old_name, rename.new_name, rename.execution_model);
-
-	auto entry_points = compiler->get_entry_points_and_stages();
-	auto entry_point = args.entry;
-	ExecutionModel model = ExecutionModelMax;
-
-	if (!args.entry_stage.empty())
-	{
-		model = stage_to_execution_model(args.entry_stage);
-		if (entry_point.empty())
-		{
-			// Just use the first entry point with this stage.
-			for (auto &e : entry_points)
-			{
-				if (e.execution_model == model)
-				{
-					entry_point = e.name;
-					break;
-				}
-			}
-
-			if (entry_point.empty())
-			{
-				fprintf(stderr, "Could not find an entry point with stage: %s\n", args.entry_stage.c_str());
-				return EXIT_FAILURE;
-			}
-		}
-		else
-		{
-			// Make sure both stage and name exists.
-			bool exists = false;
-			for (auto &e : entry_points)
-			{
-				if (e.execution_model == model && e.name == entry_point)
-				{
-					exists = true;
-					break;
-				}
-			}
-
-			if (!exists)
-			{
-				fprintf(stderr, "Could not find an entry point %s with stage: %s\n", entry_point.c_str(),
-				        args.entry_stage.c_str());
-				return EXIT_FAILURE;
-			}
-		}
-	}
-	else if (!entry_point.empty())
-	{
-		// Make sure there is just one entry point with this name, or the stage
-		// is ambiguous.
-		uint32_t stage_count = 0;
-		for (auto &e : entry_points)
-		{
-			if (e.name == entry_point)
-			{
-				stage_count++;
-				model = e.execution_model;
-			}
-		}
-
-		if (stage_count == 0)
-		{
-			fprintf(stderr, "There is no entry point with name: %s\n", entry_point.c_str());
-			return EXIT_FAILURE;
-		}
-		else if (stage_count > 1)
-		{
-			fprintf(stderr, "There is more than one entry point with name: %s. Use --stage.\n", entry_point.c_str());
-			return EXIT_FAILURE;
-		}
-	}
-
-	if (!entry_point.empty())
-		compiler->set_entry_point(entry_point, model);
-
-	if (!args.set_version && !compiler->get_common_options().version)
-	{
-		fprintf(stderr, "Didn't specify GLSL version and SPIR-V did not specify language.\n");
-		print_help();
-		return EXIT_FAILURE;
-	}
-
-	CompilerGLSL::Options opts = compiler->get_common_options();
-	if (args.set_version)
-		opts.version = args.version;
-	if (args.set_es)
-		opts.es = args.es;
-	opts.force_temporary = args.force_temporary;
-	opts.separate_shader_objects = args.sso;
-	opts.flatten_multidimensional_arrays = args.flatten_multidimensional_arrays;
-	opts.enable_420pack_extension = args.use_420pack_extension;
-	opts.vulkan_semantics = args.vulkan_semantics;
-	opts.vertex.fixup_clipspace = args.fixup;
-	opts.vertex.flip_vert_y = args.yflip;
-	opts.vertex.support_nonzero_base_instance = args.support_nonzero_baseinstance;
-	opts.emit_push_constant_as_uniform_buffer = args.glsl_emit_push_constant_as_ubo;
-	compiler->set_common_options(opts);
-
-	// Set HLSL specific options.
-	if (args.hlsl)
-	{
-		auto *hlsl = static_cast<CompilerHLSL *>(compiler.get());
-		auto hlsl_opts = hlsl->get_hlsl_options();
-		if (args.set_shader_model)
-		{
-			if (args.shader_model < 30)
-			{
-				fprintf(stderr, "Shader model earlier than 30 (3.0) not supported.\n");
-				return EXIT_FAILURE;
-			}
-
-			hlsl_opts.shader_model = args.shader_model;
-		}
-
-		if (args.hlsl_compat)
-		{
-			// Enable all compat options.
-			hlsl_opts.point_size_compat = true;
-			hlsl_opts.point_coord_compat = true;
-		}
-
-		if (hlsl_opts.shader_model <= 30)
-		{
-			combined_image_samplers = true;
-			build_dummy_sampler = true;
-		}
-
-		hlsl_opts.support_nonzero_base_vertex_base_instance = args.hlsl_support_nonzero_base;
-		hlsl->set_hlsl_options(hlsl_opts);
-	}
-
-	if (build_dummy_sampler)
-	{
-		uint32_t sampler = compiler->build_dummy_sampler_for_combined_images();
-		if (sampler != 0)
-		{
-			// Set some defaults to make validation happy.
-			compiler->set_decoration(sampler, DecorationDescriptorSet, 0);
-			compiler->set_decoration(sampler, DecorationBinding, 0);
-		}
-	}
-
-	ShaderResources res;
-	if (args.remove_unused)
-	{
-		auto active = compiler->get_active_interface_variables();
-		res = compiler->get_shader_resources(active);
-		compiler->set_enabled_interface_variables(move(active));
-	}
-	else
-		res = compiler->get_shader_resources();
-
-	if (args.flatten_ubo)
-	{
-		for (auto &ubo : res.uniform_buffers)
-			compiler->flatten_buffer_block(ubo.id);
-		for (auto &ubo : res.push_constant_buffers)
-			compiler->flatten_buffer_block(ubo.id);
-	}
-
-	auto pls_inputs = remap_pls(args.pls_in, res.stage_inputs, &res.subpass_inputs);
-	auto pls_outputs = remap_pls(args.pls_out, res.stage_outputs, nullptr);
-	compiler->remap_pixel_local_storage(move(pls_inputs), move(pls_outputs));
-
-	for (auto &ext : args.extensions)
-		compiler->require_extension(ext);
-
-	for (auto &remap : args.remaps)
-	{
-		if (remap_generic(*compiler, res.stage_inputs, remap))
-			continue;
-		if (remap_generic(*compiler, res.stage_outputs, remap))
-			continue;
-		if (remap_generic(*compiler, res.subpass_inputs, remap))
-			continue;
-	}
-
-	for (auto &rename : args.interface_variable_renames)
-	{
-		if (rename.storageClass == StorageClassInput)
-			spirv_cross_util::rename_interface_variable(*compiler, res.stage_inputs, rename.location,
-			                                            rename.variable_name);
-		else if (rename.storageClass == StorageClassOutput)
-			spirv_cross_util::rename_interface_variable(*compiler, res.stage_outputs, rename.location,
-			                                            rename.variable_name);
-		else
-		{
-			fprintf(stderr, "error at --rename-interface-variable <in|out> ...\n");
-			return EXIT_FAILURE;
-		}
-	}
-
-	if (args.dump_resources)
-	{
-		print_resources(*compiler, res);
-		print_push_constant_resources(*compiler, res.push_constant_buffers);
-		print_spec_constants(*compiler);
-		print_capabilities_and_extensions(*compiler);
-	}
-
-	if (combined_image_samplers)
-	{
-		compiler->build_combined_image_samplers();
-		if (args.combined_samplers_inherit_bindings)
-			spirv_cross_util::inherit_combined_sampler_bindings(*compiler);
-
-		// Give the remapped combined samplers new names.
-		for (auto &remap : compiler->get_combined_image_samplers())
-		{
-			compiler->set_name(remap.combined_id, join("SPIRV_Cross_Combined", compiler->get_name(remap.image_id),
-			                                           compiler->get_name(remap.sampler_id)));
-		}
-	}
-
-	if (args.hlsl)
-	{
-		auto *hlsl_compiler = static_cast<CompilerHLSL *>(compiler.get());
-		uint32_t new_builtin = hlsl_compiler->remap_num_workgroups_builtin();
-		if (new_builtin)
-		{
-			hlsl_compiler->set_decoration(new_builtin, DecorationDescriptorSet, 0);
-			hlsl_compiler->set_decoration(new_builtin, DecorationBinding, 0);
-		}
-	}
-
-	string glsl;
-	for (uint32_t i = 0; i < args.iterations; i++)
-	{
-		if (args.hlsl)
-		{
-			for (auto &remap : args.hlsl_attr_remap)
-				static_cast<CompilerHLSL *>(compiler.get())->add_vertex_attribute_remap(remap);
-		}
-
-		glsl = compiler->compile();
+		for (unsigned i = 0; i < args.iterations; i++)
+			compiled_output = compile_iteration(args, spirv_file);
 	}
 
 	if (args.output)
-		write_string_to_file(args.output, glsl.c_str());
+		write_string_to_file(args.output, compiled_output.c_str());
 	else
-		printf("%s", glsl.c_str());
+		printf("%s", compiled_output.c_str());
 
 	return EXIT_SUCCESS;
 }