--- a/spirv_cfg.cpp
+++ b/spirv_cfg.cpp
@@ -143,7 +143,7 @@ void CFG::build_post_order_visit_order()
 
 void CFG::add_branch(uint32_t from, uint32_t to)
 {
-	const auto add_unique = [](vector<uint32_t> &l, uint32_t value) {
+	const auto add_unique = [](SmallVector<uint32_t> &l, uint32_t value) {
 		auto itr = find(begin(l), end(l), value);
 		if (itr == end(l))
 			l.push_back(value);
@@ -223,4 +223,4 @@ void DominatorBuilder::lift_continue_block_dominator()
 	if (back_edge_dominator)
 		dominator = cfg.get_function().entry_block;
 }
-} // namespace spirv_cross
+} // namespace SPIRV_CROSS_NAMESPACE