SPIRV-Cross is rather heavy on global allocations with the use of data structures required to implement the codegen. There are many, many opportunities to pool them, avoid them, allocate more on stack, etc.

This PR is about reducing global allocations as much as is reasonable. It is not about rewriting lots of code significantly, but replace data structures where it makes sense.

- Replace ostringstream with custom implementation. This is the biggest win.
  ~30% performance uplift on vector-shuffle-oom test.
  Allocations are measurably reduced in Valgrind.
    
- Replace std::vector with SmallVector.
  Classic malloc optimization, small vectors are backed by inline data.
  ~ 7-8% gain on vector-shuffle-oom on GCC 8 on Linux.
    
- Use an object pool for IVariant type.
  We generally allocate a lot of SPIR* objects. We can amortize these
  allocations neatly by pooling them.

- Add a way to pass down --iterations down to spirv-cross in test_shaders.py. This allows for large performance sweeps over a shader test set.
    
- ~15% overall uplift on ./test_shaders.py --iterations 10000 shaders/. Which has lots of other overhead and tiny shaders.

I spent some time on trying to do similar optimizations for unordered_map and friends but I failed to implement a working std::allocator replacement. The only way forward here is to write custom hashmap implementations I think, but that is a lot of work we can punt.

Fix #923.