@HansKristian-Work 2019-04-09
@billhollings @cdavis5e This should stave off a lot of the overhead. I tried building this commit in MoltenVK without issue. If you have some performance data points from real games, that'd be nice to see!