@billhollings 2019-04-09
@HansKristian-Work 

Thanks for all the effort on this!

MoltenVK has the `MVK_CONFIG_PERFORMANCE_TRACKING` env var...which if run with a Debug build of MoltenVK...will cause shader conversion performance statistics to be logged (I should change the Debug build requirement for logging to an env var too at some point). With a Release build...the performance stats are not logged...but are queryable through the API.

I'll try running a comparison against Dota2 when I get a chance.