@billhollings 2019-04-10
@HansKristian-Work 

Hmmm...in Dota2 I am not seeing any appreciable change in the average conversion time across 350 shaders. 

What kind of numbers are you seeing...in other words...what should I expect to see? I'm suspicious that I'm getting more or less identical results between Before vs After. Perhaps I have screwed up my testing.

And a live game is pretty noisy (as is evident from the variations across runs from a single build) and may not be an accurate measure of numerical improvements.

The values below are 3 separate runs of Dota2 startup for each of Before and After...and include just the SPIRV-Cross conversions (no loading or MSL compilation). Each run converts about 350 shaders.

BEFORE:

```
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 4.818 ms, avg: 3.212 ms, min: 0.304 ms, max: 33.480 ms, count: 335
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 4.443 ms, avg: 5.116 ms, min: 0.300 ms, max: 94.172 ms, count: 344
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 5.518 ms, avg: 5.602 ms, min: 0.395 ms, max: 114.684 ms, count: 347
```

AFTER:

```
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 3.826 ms, avg: 5.436 ms, min: 0.448 ms, max: 162.127 ms, count: 350
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 3.748 ms, avg: 3.593 ms, min: 0.384 ms, max: 99.548 ms, count: 340
[mvk-info] Performance to convert SPIR-V to MSL source code curr: 3.793 ms, avg: 5.653 ms, min: 0.505 ms, max: 110.268 ms, count: 348
```