@HansKristian-Work 2019-04-10
Thanks for testing. That's very erratic performance, doesn't seem like we can draw any conclusions. I'll try to consider how we can isolate this from the application with a specialized Fossilize replayer perhaps ... That max timing value looks very suspicious.