@billhollings 2019-04-10
Yeah...there is a lot of multithreading going on in shader and pipeline prep in Dota2. So I suspect the max values are outliers that get caught being delayed because of thread starvation.

I could probably add similar timing tracking to the MoltenVKShaderConverter tool and run it across an entire directory tree. Or something similar could be added to `spirv-cross main` to do the same?